+++ SleepHunter/Services/Clients/IClientLaunchInteraction.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace SleepHunter.Services.Clients
+{
+    public interface IClientLaunchInteraction
+    {
+        void ShowError(Exception exception);
+    }
+}